Resorting file structure (Tidying)
Polishes that are general
Resize the image
