+++ brainwallet/Additions/CGRect+Additions.swift
