+++ brainwallet/Additions/NumberFormatter+Additions.swift
