+++ brainwallet/Additions/String+Additions.swift
