+++ brainwallet/Additions/UIBarButtonItem+Additions.swift
