+++ brainwallet/Additions/UIView+AnimationAdditions.swift
