+++ brainwallet/Additions/UIView+InitAdditions.swift
