+++ brainwallet/Additions/UIViewController+BRWAdditions.swift
