+++ brainwallet/CoreResources/BIP39Words.plist
