+++ brainwallet/CoreResources/module.modulemap
