+++ brainwallet/Extensions/NSNotificationNameExtension.swift
