+++ brainwallet/Extensions/TextView+Extension.swift
