+++ brainwallet/Extensions/UIApplication+Extension.swift
