+++ brainwallet/Extensions/UnsafeMutablePointerExtension.swift
