+++ brainwallet/PinPadViewController.swift
@@ -195,212 +195,3 @@ class PinPadViewController: UICollectionViewController {
 		fatalError("init(coder:) has not been implemented")
 	}
 }
-
-class GenericPinPadCell: UICollectionViewCell {
-	var text: String? {
-		didSet {
-			if text == deleteKeyIdentifier {
-				imageView.image = #imageLiteral(resourceName: "Delete")
-				topLabel.text = ""
-				centerLabel.text = ""
-			} else {
-				imageView.image = nil
-				topLabel.text = text
-				centerLabel.text = text
-			}
-			setAppearance()
-			setSublabel()
-		}
-	}
-
-	let sublabels = [
-		"2": "",
-		"3": "",
-		"4": "",
-		"5": "",
-		"6": "",
-		"7": "",
-		"8": "",
-		"9": "",
-	]
-
-	override var isHighlighted: Bool {
-		didSet {
-			guard text != "" else { return } // We don't want the blank cell to highlight
-			setAppearance()
-		}
-	}
-
-	override init(frame: CGRect) {
-		super.init(frame: frame)
-		setup()
-	}
-
-	let topLabel = UILabel(font: .customBody(size: 28.0))
-	let centerLabel = UILabel(font: .customBody(size: 28.0))
-	let sublabel = UILabel(font: .customBody(size: 11.0))
-	let imageView = UIImageView()
-
-	private func setup() {
-		setAppearance()
-		topLabel.textAlignment = .center
-		centerLabel.textAlignment = .center
-		sublabel.textAlignment = .center
-		addSubview(topLabel)
-		addSubview(centerLabel)
-		addSubview(sublabel)
-		addSubview(imageView)
-		imageView.contentMode = .center
-		addConstraints()
-	}
-
-	func addConstraints() {
-		imageView.constrain(toSuperviewEdges: nil)
-		centerLabel.constrain(toSuperviewEdges: nil)
-		topLabel.constrain([
-			topLabel.centerXAnchor.constraint(equalTo: centerXAnchor),
-			topLabel.topAnchor.constraint(equalTo: topAnchor, constant: 2.5),
-		])
-		sublabel.constrain([
-			sublabel.centerXAnchor.constraint(equalTo: centerXAnchor),
-			sublabel.topAnchor.constraint(equalTo: topLabel.bottomAnchor, constant: -3.0),
-		])
-	}
-
-	override var isAccessibilityElement: Bool {
-		get {
-			return true
-		}
-		set {}
-	}
-
-	override var accessibilityLabel: String? {
-		get {
-			return topLabel.text
-		}
-		set {}
-	}
-
-	override var accessibilityTraits: UIAccessibilityTraits {
-		get {
-			return UIAccessibilityTraits.staticText
-		}
-		set {}
-	}
-
-	func setAppearance() {}
-	func setSublabel() {}
-
-	@available(*, unavailable)
-	required init?(coder _: NSCoder) {
-		fatalError("init(coder:) has not been implemented")
-	}
-}
-
-class ClearNumberPad: GenericPinPadCell {
-	override func setAppearance() {
-		if text == "0" {
-			topLabel.isHidden = true
-			centerLabel.isHidden = false
-		} else {
-			topLabel.isHidden = false
-			centerLabel.isHidden = true
-		}
-
-        topLabel.textColor = BrainwalletUIColor.content
-		centerLabel.textColor = BrainwalletUIColor.content
-		sublabel.textColor = BrainwalletUIColor.content
-
-		if isHighlighted {
-            backgroundColor = BrainwalletUIColor.gray.withAlphaComponent(0.2)
-		} else {
-			backgroundColor = BrainwalletUIColor.surface
-
-			if text == "" || text == deleteKeyIdentifier {
-				imageView.image = imageView.image?.withRenderingMode(.alwaysTemplate)
-                imageView.tintColor = BrainwalletUIColor.content
-			}
-		}
-	}
-
-	override func setSublabel() {
-		guard let text = text else { return }
-		if sublabels[text] != nil {
-			sublabel.text = sublabels[text]
-		}
-	}
-}
-
-class ClearDecimalPad: GenericPinPadCell {
-	override func setAppearance() {
-		centerLabel.backgroundColor = .clear
-		imageView.image = imageView.image?.withRenderingMode(.alwaysTemplate)
-
-		if isHighlighted {
-			centerLabel.textColor = BrainwalletUIColor.content
-            imageView.tintColor = BrainwalletUIColor.content
-		} else {
-            centerLabel.textColor = BrainwalletUIColor.surface
-			imageView.tintColor = BrainwalletUIColor.surface
-		}
-	}
-
-	override func addConstraints() {
-		centerLabel.constrain(toSuperviewEdges: nil)
-		imageView.constrain(toSuperviewEdges: nil)
-	}
-}
-
-class WhiteDecimalPad: GenericPinPadCell {
-	override func setAppearance() {
-		if isHighlighted {
-            centerLabel.backgroundColor = BrainwalletUIColor.surface
-			centerLabel.textColor = BrainwalletUIColor.content
-		} else {
-            centerLabel.backgroundColor = BrainwalletUIColor.surface
-			centerLabel.textColor = BrainwalletUIColor.content
-		}
-	}
-
-	override func addConstraints() {
-		centerLabel.constrain(toSuperviewEdges: nil)
-		imageView.constrain(toSuperviewEdges: nil)
-	}
-}
-
-class WhiteNumberPad: GenericPinPadCell {
-	override func setAppearance() {
-		if text == "0" {
-			topLabel.isHidden = true
-			centerLabel.isHidden = false
-		} else {
-			topLabel.isHidden = false
-			centerLabel.isHidden = true
-		}
-
-		if isHighlighted {
-            backgroundColor = BrainwalletUIColor.surface
-			topLabel.textColor = BrainwalletUIColor.content
-			centerLabel.textColor = BrainwalletUIColor.content
-			sublabel.textColor = BrainwalletUIColor.content
-		} else {
-			if text == "" || text == deleteKeyIdentifier {
-                backgroundColor = BrainwalletUIColor.surface
-				imageView.image = imageView.image?.withRenderingMode(.alwaysTemplate)
-                imageView.tintColor = BrainwalletUIColor.content
-			} else {
-                backgroundColor = BrainwalletUIColor.surface
-				topLabel.textColor = BrainwalletUIColor.content
-				centerLabel.textColor = BrainwalletUIColor.content
-				sublabel.textColor = BrainwalletUIColor.content
-			}
-		}
-	}
-
-	override func setSublabel() {
-		guard let text = text else { return }
-		if sublabels[text] != nil {
-			sublabel.text = sublabels[text]
-		}
-	}
-}
